requirements.txt
ultralytics
easyocr
opencv-python
google-generativeai
gtts
python-dotenv
thefuzz
